VERSION 0.0.0.2
The SpecPlotter function now extracts the directory path, file name, and file extension from a single character string. It also provides options to save the spectral data to an rspec object in the global environment, save an image of the resulting plot alongside the file of interest, and control some visualization parameters.